feat: added global error handler
